Move retryable error to errors package
