+++ pkg/runtime/runtime_test.go
@@ -3185,7 +3185,7 @@ func TestOnNewPublishedMessage(t *testing.T) {
 		// assert
 		var cloudEvent map[string]interface{}
 		json.Unmarshal(testPubSubMessage.data, &cloudEvent)
-		expectedClientError := fmt.Errorf("RETRY status returned from app while processing pub/sub event %v: %w", cloudEvent["id"].(string), NewRetriableError(nil))
+		expectedClientError := fmt.Errorf("RETRY status returned from app while processing pub/sub event %v: %w", cloudEvent["id"].(string), rterrors.NewRetriable(nil))
 		assert.Equal(t, expectedClientError.Error(), err.Error())
 		mockAppChannel.AssertNumberOfCalls(t, "InvokeMethod", 1)
 	})
@@ -3281,7 +3281,7 @@ func TestOnNewPublishedMessage(t *testing.T) {
 		err := rt.publishMessageHTTP(context.Background(), testPubSubMessage)
 
 		// assert
-		expectedError := fmt.Errorf("error returned from app channel while sending pub/sub event to app: %w", NewRetriableError(invokeError))
+		expectedError := fmt.Errorf("error returned from app channel while sending pub/sub event to app: %w", rterrors.NewRetriable(invokeError))
 		assert.Equal(t, expectedError.Error(), err.Error(), "expected errors to match")
 		mockAppChannel.AssertNumberOfCalls(t, "InvokeMethod", 1)
 	})
@@ -3323,7 +3323,7 @@ func TestOnNewPublishedMessage(t *testing.T) {
 		var cloudEvent map[string]interface{}
 		json.Unmarshal(testPubSubMessage.data, &cloudEvent)
 		errMsg := fmt.Sprintf("retriable error returned from app while processing pub/sub event %v, topic: %v, body: Internal Error. status code returned: 500", cloudEvent["id"].(string), cloudEvent["topic"])
-		expectedClientError := NewRetriableError(errors.New(errMsg))
+		expectedClientError := rterrors.NewRetriable(errors.New(errMsg))
 		assert.Equal(t, expectedClientError.Error(), err.Error())
 		mockAppChannel.AssertNumberOfCalls(t, "InvokeMethod", 1)
 	})
@@ -3395,7 +3395,7 @@ func TestOnNewPublishedMessageGRPC(t *testing.T) {
 			expectedError: fmt.Errorf(
 				"error returned from app while processing pub/sub event %v: %w",
 				testPubSubMessage.cloudEvent[pubsub.IDField],
-				NewRetriableError(status.Error(codes.Unknown, assert.AnError.Error())),
+				rterrors.NewRetriable(status.Error(codes.Unknown, assert.AnError.Error())),
 			),
 		},
 		{
@@ -3420,7 +3420,7 @@ func TestOnNewPublishedMessageGRPC(t *testing.T) {
 			expectedError: fmt.Errorf(
 				"RETRY status returned from app while processing pub/sub event %v: %w",
 				testPubSubMessage.cloudEvent[pubsub.IDField],
-				NewRetriableError(nil),
+				rterrors.NewRetriable(nil),
 			),
 		},
 		{
@@ -3436,7 +3436,7 @@ func TestOnNewPublishedMessageGRPC(t *testing.T) {
 				"unknown status returned from app while processing pub/sub event %v, status: %v, err: %w",
 				testPubSubMessage.cloudEvent[pubsub.IDField],
 				runtimev1pb.TopicEventResponse_TopicEventResponseStatus(99),
-				NewRetriableError(nil),
+				rterrors.NewRetriable(nil),
 			),
 		},
 		{
